API Version2
This version of the Creates a FieldCommunicator that understands how
to send/receive field data.

Current implimentation uses a single mesh partition. This change removes the ability to create a coupler with multiple partitions. The use case for multiple partitioning schemes is not well understood at theis point and supporting multiple partitions adds extra API complexity for the user trying to set up a coupling scheme. The envisioned use case for multiple partitions would be a coupling server that is simultaneously managing multiple coupling schemes simultaneously. There are many subtleties to this use case and it would be best to have a cohesive methodology for supporting this when/if the need arises.