mdlent owners set w/ equal order classification
In Omega_h all the mesh entities classified on a given part boundary
don't necessarily have the same ownership.
For example, in a 2d mesh, the mesh vertices on a partition model edge
bounding two partition model faces (parts), A and B, may be owned by
rank A while the mesh edges classified on that edge can be owned by rank
B.
The code that was deriving the ownership did not account for this and is
now fixed by assigning the owner of a model entity as the:
minimum(owner(meshEnt) for all meshEnt classified on modelEnt) where
meshEnt and modelEnt have the same dimension and owner(meshEnt) returns
the rank of the process that owns that mesh entity.
The ownership of mesh entities is consistent across all procesess with a
copy of a given mesh entity, thus this model entity ownership protocol
does not require additional communication.

Note, this commit passes the ctest `test_proxy_coupling` test using four
rendezvous ranks and gets past the assertion reported in #114 for a new
16 rank rendezvous test case (`test_proxy_coupling_40p`) but later fails with

```
local_gids.size() == received_gids.size() failed at
/space/cwsmith/wdmCpl/wdmapp_coupling/src/wdmcpl.h + 78
```