Outline of new coupling API including field trasfer
This is a rough outline of the coupling API that includes field transfer
operations. Currently not all components of this API are functional and
it doesn't fully compile. However, the test_proxy_coupling includes the
desired API usage. The field shim/omega_h field shim should be set.
Some minor API fixes in the Coupler class need to be made.

Major work that's still needed:
- Grid search on GPU
- Consistency of MemorySpaces throughout application (to enable run on
  GPU)
- Using typed coordinate value types