Gather/Scatter "running"
Everything runs without issue. Unfortunately, it seems as though there
is still a minor data corruption where a few of the points are
imporperly getting set. It seems most likely that this is due to the
filtering and use of the "set" function possibly due to the calculation
of the filter mask.