Update to dealing with XGC fields
Relies on REDEV: 1a157d0fbe95485ab698b09aa52d220d49bba6a9

XGC fields have two issues that make them challenging to deal with.
First, Adios2 communication does not happen on all of the ranks within
the global communicator, but some operations in the XGC field must take
place over the global communicator. This requires that the RDV Comm should be able to
handle a communicator that is NULL on some ranks.

Second issue is that the XGC fields do not span the entire global
communicator, but we must add all fields across the global communicator
because some operations in the field setup phase are global across the
client communication communicator.