move the coordinates of the DOFHolders to the FieldLayout
As I'm working through the field updates it has become clear that the DOF
coordinates should be part of the field layout. This becomes clear when
you realize that computing the ReversePartitionMap requires the coordinates
and the reverse partition map is a key component of the FieldLayout.
To a large extent, Field's only responsibility should be to know how to do
evaluations. Likely this indicates that the Serialize/Deserialize functionality
should be pushed down to the FieldLayout as well.